+++ src/MOLECULE/pair_tip4p_cut.cpp
@@ -550,8 +550,8 @@ void PairTIP4PCut::compute_newsite(double *xO,  double *xH1,
 
 double PairTIP4PCut::memory_usage()
 {
-  double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += 2 * nmax * sizeof(double);
+  double bytes = (double)maxeatom * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)2 * nmax * sizeof(double);
   return bytes;
 }
