+++ src/USER-SMD/compute_smd_triangle_vertices.cpp
@@ -118,6 +118,6 @@ void ComputeSMDTriangleVertices::compute_peratom() {
  ------------------------------------------------------------------------- */
 
 double ComputeSMDTriangleVertices::memory_usage() {
-    double bytes = size_peratom_cols * nmax * sizeof(double);
+    double bytes = (double)size_peratom_cols * nmax * sizeof(double);
     return bytes;
 }
