+++ src/USER-SMD/compute_smd_ulsph_strain_rate.cpp
@@ -119,6 +119,6 @@ void ComputeSMDULSPHStrainRate::compute_peratom() {
  ------------------------------------------------------------------------- */
 
 double ComputeSMDULSPHStrainRate::memory_usage() {
-        double bytes = size_peratom_cols * nmax * sizeof(double);
+        double bytes = (double)size_peratom_cols * nmax * sizeof(double);
         return bytes;
 }
